+++ src/metabrowser/builtin_plugins/folder/index.js
@@ -1,8 +1,7 @@
 // Built-in folder view and Overview contribution registration.
 
 import { createCategoryPalettePool } from "./category_palette.js";
-import { createFileTotalsPanel } from "./file_totals_panel.js";
-import { createFileTypeSummaryPanel } from "./file_type_summary.js";
+import { createFileOverviewPanel } from "./file_overview_panel.js";
 import { createOverviewView } from "./overview.js";
 import { createFolderOverviewRegistry } from "./overview_registry.js";
 import { createReadmePanel } from "./readme_panel.js";
@@ -23,12 +22,8 @@ const rollupControls = createFolderRollupControls(mb);
 const rollupProjectionPool = createFolderRollupProjectionPool();
 mb.folderOverview = registry;
 registry.registerPanel(
-  "folder.file-totals",
-  createFileTotalsPanel(mb, palettePool, rollupProjectionPool, rollupControls),
-);
-registry.registerPanel(
-  "folder.file-types",
-  createFileTypeSummaryPanel(mb, palettePool, rollupProjectionPool, rollupControls),
+  "folder.file-overview",
+  createFileOverviewPanel(mb, palettePool, rollupProjectionPool, rollupControls),
 );
 registry.registerPanel("folder.readme", createReadmePanel(mb));
 mb.registerView("folder", "overview", createOverviewView(mb, registry));